refactor: Piece 추상클래스 하위에 연속이동기물과 제한된이동기물 분리
