+++ src/main/java/chess/domain/piece/Bishop.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.result.Score;
 
-public class Bishop extends Piece {
+public class Bishop extends ConsecutiveMovablePiece {
     private static final String NAME = "B";
     private static final double SCORE = 3;
