+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,24 +1,12 @@
 package chess.domain.piece;
 
-import chess.domain.board.ChessBoard;
-import chess.domain.board.Coordinate;
 import chess.domain.result.Score;
 
-public class Knight extends Piece {
+public class Knight extends LimitedMovablePiece {
     private static final String NAME = "N";
     private static final double SCORE = 2.5;
 
     public Knight(TeamType teamType) {
         super(NAME, teamType, new Score(SCORE), Direction.getKnightDirections());
-    }
-
-    @Override
-    public boolean isMovable(ChessBoard chessBoard, Coordinate current, Coordinate destination) {
-        Direction direction = current.evaluateDirection(destination);
-        if (!isCorrectDirection(direction)) {
-            return false;
-        }
-        Coordinate nextCoordinate = current.move(direction);
-        return nextCoordinate.equals(destination) && chessBoard.isEmptyOrHasEnemyOn(destination, getTeamType());
     }
 }
