+++ src/main/java/chess/domain/piece/Queen.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.result.Score;
 
-public class Queen extends Piece {
+public class Queen extends ConsecutiveMovablePiece {
     private static final String NAME = "Q";
     private static final double SCORE = 9;
