+++ src/main/java/chess/domain/piece/Rook.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.result.Score;
 
-public class Rook extends Piece {
+public class Rook extends ConsecutiveMovablePiece {
     private static final String NAME = "R";
     private static final double SCORE = 5;
